TE-10017 Fix IdeHelper typehints, upgrade PHPStan levels. (#8792)
* TE-10017 Fix IdeHelper typehints, upgrade PHPStan levels.

* TE-10017 Fix CS

* TE-10017 Remove unneeded ignores.

* TE-10017 Fix IdeHelper typehints, upgrade PHPStan levels.

* TE-10017 Fix IdeHelper typehints, upgrade PHPStan levels.

* TE-10017 Fix IdeHelper typehints, upgrade PHPStan levels.

* TE-10017 Fix IdeHelper typehints, upgrade PHPStan levels.

* TE-10017 Adding now found missing transfer declarations.

+++ Bundles/SalesStatistics/src/Spryker/Zed/SalesStatistics/Persistence/SalesStatisticsRepository.php
@@ -51,7 +51,7 @@ class SalesStatisticsRepository extends AbstractRepository implements SalesStati
      */
     public function getOrderCountStatisticByDays(int $day): ChartDataTraceTransfer
     {
-        $date = date('Y-m-d H:i:s.u', strtotime(sprintf('-%d days', $day)));
+        $date = date('Y-m-d H:i:s.u', (int)strtotime(sprintf('-%d days', $day)));
 
         $result = $this->getDataOrderCountStatisticByDays($date);
